feat: ScriptInfo数据类型更改 - ver_num、last_ver、download_count更改为u32 - 简化脚本编辑器保存脚本逻辑

+++ src-tauri/crates/runtime_engine/src/domain/scripts/script_info.rs
@@ -99,9 +99,9 @@ pub struct ScriptInfo {
     pub create_time: Option<String>,
     pub update_time: Option<String>,
     pub ver_name: String,
-    pub ver_num: u64,
-    pub latest_ver: u64,
-    pub download_count: u64,
+    pub ver_num: u32,
+    pub latest_ver: u32,
+    pub download_count: u32,
     pub script_type: ScriptType,
     #[serde(default)]
     pub is_valid: bool,
